Refactor: Lift state to App component
